+++ src/components/index.ts
@@ -1,4 +1,4 @@
-export { Button } from "./Button/button";
+export { Button } from "./button/button";
 export { Card } from "./card/card";
 export { Game } from "./game/game";
-export { TextField } from "./TextField/TextField";
+export { TextField } from "./text-field/text-field";
